requirements.txt
openai>=1.46,<2
python-dotenv>=1.0
gradio==4.44.0
gradio_client==1.3.0
fastapi==0.115.2
starlette==0.40.0
anyio==4.4.0
aiofiles==23.2.1
ffmpy==0.3.2
pydantic==2.10.6
pypdf>=5.0

